+++ Changes.md
@@ -60,6 +60,8 @@ API
   - `key.setTime()` will not remove active clashing key, instead it becomes become inactive. inactive key at old time becomes active.
   - Added `curve.removeInactiveKeys()` function, removes all the inactive keys parented to a curve.
   - Added `key.isActive()` function, returns true if key is parented to curve and is the active key at its current time.
+- AnimationGadget :
+  - Added `selectedKeys()` function, returns current set of selected keys.
 
 Breaking Changes
 ----------------